Nvim: consolidates telescope files
